requirements.txt
python-telegram-bot>=20.0
requests>=2.25.1
python-dotenv>=1.0.0
anthropic>=0.3.11
